Allow os-access-token-type for device auth
RFC8628 is part of OAuth, which doesn't standardize ID tokens. EntraID
does provide an id_token in the token response (although not all IDPs
do).

Depending on the Keystone and EntraID configurations, an id_token might
pass introspection while an access_token doesn't.

This allows a user to select which token they would like to use when
requesting a keystone token.

Change-Id: Idcddd8b58d862968b6412c87420a44525d96ce91
Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>